refactor library info form into its own component, removing repetitive code in library creation and editing components
